backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase==2.3.0
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
slowapi==0.1.9
python-jose[cryptography]==3.3.0
bleach==6.1.0
